# Histogram performance: optimize floatBucketIterator