There are two optimizations in this PR:
1. Introduce `strippedBucket` struct/`strippedAt` method on `floatBucketIterator` which allows to skip bound calculations (which are relatively expensive `getBound` calls) in use cases where bound values are not needed (such as `DetectReset`).
2. Avoid allocating `floatBucketIterator` object on heap.

For this change, I have adapted benchmark code from https://github.com/prometheus/prometheus/pull/12711/commits/4787c879bce2b44a8a5634d94f1c8b847dc07875. Benchmark results below.

Before the change:
```
go test -bench=BenchmarkFloatHistogramDetectReset -count=5             
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/histogram
BenchmarkFloatHistogramDetectReset-10    	 9413133	       126.1 ns/op	     256 B/op	       2 allocs/op
BenchmarkFloatHistogramDetectReset-10    	 9557922	       124.6 ns/op	     256 B/op	       2 allocs/op
BenchmarkFloatHistogramDetectReset-10    	 9350709	       128.9 ns/op	     256 B/op	       2 allocs/op
BenchmarkFloatHistogramDetectReset-10    	 9439388	       128.7 ns/op	     256 B/op	       2 allocs/op
BenchmarkFloatHistogramDetectReset-10    	 9173514	       127.0 ns/op	     256 B/op	       2 allocs/op
```

Avoiding `getBound` calls (~35% improvement):
```
BenchmarkFloatHistogramDetectReset-10    	14278440	        81.12 ns/op	     256 B/op	       2 allocs/op
BenchmarkFloatHistogramDetectReset-10    	13361082	        82.16 ns/op	     256 B/op	       2 allocs/op
BenchmarkFloatHistogramDetectReset-10    	15049930	        80.98 ns/op	     256 B/op	       2 allocs/op
BenchmarkFloatHistogramDetectReset-10    	14484721	        83.29 ns/op	     256 B/op	       2 allocs/op
BenchmarkFloatHistogramDetectReset-10    	14899450	        83.53 ns/op	     256 B/op	       2 allocs/op
```

Additionally, avoiding iterator allocations (~2.5x total improvement, allocs down from 2 to 0):
```
BenchmarkFloatHistogramDetectReset-10    	24134176	        49.68 ns/op	       0 B/op	       0 allocs/op
BenchmarkFloatHistogramDetectReset-10    	23310399	        49.72 ns/op	       0 B/op	       0 allocs/op
BenchmarkFloatHistogramDetectReset-10    	24175940	        49.61 ns/op	       0 B/op	       0 allocs/op
BenchmarkFloatHistogramDetectReset-10    	23970396	        49.60 ns/op	       0 B/op	       0 allocs/op
BenchmarkFloatHistogramDetectReset-10    	23984430	        49.62 ns/op	       0 B/op	       0 allocs/op
```

As a followup to this PR, I am planning to eliminate (or reduce) `getBound` calls from `floatBucketIterator.Next()` as well, which would have even broader impact in terms of use cases affected. At a quick glance, this seems to be possible because of the properties of this function (especially in case of `absoluteStartValue == 0`, which appears to be the case in query engine).