@bboreham 2023-10-06
You may find it useful to use [`benchstat`](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) to aggregate your results.
(install with `go install golang.org/x/perf/cmd/benchstat@latest`)