@linasm 2023-10-09
> There are two optimizations in this PR:
> 
> 1. Introduce `strippedBucket` struct/`strippedAt` method on `floatBucketIterator` which allows to skip bound calculations (which are relatively expensive `getBound` calls) in use cases where bound values are not needed (such as `DetectReset`).
> 2. Avoid allocating `floatBucketIterator` object on heap.

Thanks, will make use of it.