@linasm 2023-10-09
Closing in favour of more comprehensive and useful https://github.com/prometheus/prometheus/pull/12954.